requirements.txt
fastapi
uvicorn
redis
aiohttp
celery
requests
pydantic
PyCryptodome
xmltodict
python-dotenv